Rename TEX case to ver-1p and address review comments
