+++ doc/content/bib/tmap8.bib
@@ -920,15 +920,91 @@ author = {Prankul Middha and Benhui Yang and Hai Wang}
 
 
 @article{fuerst2023parametric,
-author = {Thomas F. Fuerst and Matthew D. Eklund and John A. Leland and Adriaan A. Riet and Chase N. Taylor},
-title = {{Parametric Study of the Vacuum Permeator for the Tritium Extraction eXperiment}},
-journal = {Fusion Science and Technology},
-volume = {79},
-number = {8},
-pages = {1224--1234},
-year = {2023},
-publisher = {Taylor \& Francis},
-doi = {10.1080/15361055.2023.2196237},
-URL = {https://doi.org/10.1080/15361055.2023.2196237},
-eprint = {https://doi.org/10.1080/15361055.2023.2196237}
+  author = {Thomas F. Fuerst and Matthew D. Eklund and John A. Leland and Adriaan A. Riet and Chase N. Taylor},
+  title = {{Parametric Study of the Vacuum Permeator for the Tritium Extraction eXperiment}},
+  journal = {Fusion Science and Technology},
+  volume = {79},
+  number = {8},
+  pages = {1224--1234},
+  year = {2023},
+  publisher = {Taylor \& Francis},
+  doi = {10.1080/15361055.2023.2196237},
+  url = {https://doi.org/10.1080/15361055.2023.2196237}
+}
+
+@article{S.Fukada2013MG201203,
+  title = {{Clarification of Tritium Behavior in Pb-Li Blanket System}},
+  author = {S. Fukada and T. Terai and S. Konishi and K. Katayama and T. Chikada and Y. Edao and T. Muroga and M. Shimada and B. Merrill and D. K. Sze},
+  journal = {Materials Transactions},
+  volume = {54},
+  number = {4},
+  pages = {425--429},
+  year = {2013},
+  doi = {10.2320/matertrans.MG201203},
+  url = {https://doi.org/10.2320/matertrans.MG201203}
+}
+
+@article{AIELLO2006639,
+  author = {A. Aiello and A. Ciampichetti and G. Benamati},
+  title = {Determination of hydrogen solubility in lead lithium using sole device},
+  journal = {Fusion Engineering and Design},
+  volume = {81},
+  number = {1},
+  pages = {639--644},
+  year = {2006},
+  note = {Proceedings of the Seventh International Symposium on Fusion Nuclear Technology},
+  issn = {0920-3796},
+  doi = {10.1016/j.fusengdes.2005.06.364},
+  url = {https://doi.org/10.1016/j.fusengdes.2005.06.364},
+  keywords = {Lead lithium, Solubility}
+}
+
+@article{FUERST2021118949,
+  author = {Thomas F. Fuerst and Paul W. Humrickhouse and Chase N. Taylor and Masashi Shimada},
+  title = {Surface effects on deuterium permeation through vanadium membranes},
+  journal = {Journal of Membrane Science},
+  volume = {620},
+  pages = {118949},
+  year = {2021},
+  issn = {0376-7388},
+  doi = {10.1016/j.memsci.2020.118949},
+  url = {https://doi.org/10.1016/j.memsci.2020.118949},
+  keywords = {Hydrogen isotopes, Vanadium, Palladium, Fusion technology, Dense metal membranes}
+}
+
+@incollection{VOLKL1975231,
+  author = {J. V{\"o}lkl and G. Alefeld},
+  title = {Hydrogen Diffusion in Metals},
+  editor = {A. S. Nowick and J. J. Burton},
+  booktitle = {Diffusion in Solids},
+  publisher = {Academic Press},
+  pages = {231--302},
+  year = {1975},
+  isbn = {978-0-12-522660-8},
+  doi = {10.1016/B978-0-12-522660-8.50010-6},
+  url = {https://doi.org/10.1016/B978-0-12-522660-8.50010-6}
+}
+
+@article{10.1021/j100723a033,
+  author = {Ewald Veleckis and Russell Keith Edwards},
+  title = {Thermodynamic Properties in the Systems Vanadium-Hydrogen, Niobium-Hydrogen, and Tantalum-Hydrogen},
+  journal = {The Journal of Physical Chemistry},
+  volume = {73},
+  number = {3},
+  pages = {683--692},
+  year = {1969},
+  month = {03},
+  issn = {0022-3654},
+  doi = {10.1021/j100723a033},
+  url = {https://doi.org/10.1021/j100723a033}
+}
+
+@article{linton1950mass,
+  author = {Linton, Jr., W. H. and Sherwood, T. K.},
+  title = {Mass Transfer from Solid Shapes to Water in Streamline and Turbulent Flow},
+  journal = {Chemical Engineering Progress},
+  volume = {46},
+  number = {5},
+  pages = {258--264},
+  year = {1950}
 }
