--- doc/content/verification_and_validation/figures/comparison_tex.py
@@ -1,1 +1,0 @@
-../../../../test/tests/TEX/comparison_tex.py
